Function Index

Function Index

Functions

ccall, func(x), T, for, and while.

Mod.func — Method.

func(x)

T

source

Mod.T — Type.

T

func(x)

source

ccall — Keyword.

ccall((symbol, library) or function_pointer, ReturnType, (ArgumentType1, ...), ArgumentValue1, ...)

Call function in C-exported shared library, specified by (function name, library) tuple, where each component is a string or symbol.

Note that the argument type tuple must be a literal tuple, and not a tuple-valued variable or expression. Alternatively, ccall may also be used to call a function pointer, such as one returned by dlsym.

Each ArgumentValue to the ccall will be converted to the corresponding ArgumentType, by automatic insertion of calls to unsafe_convert(ArgumentType, cconvert(ArgumentType, ArgumentValue)). (See also the documentation for each of these functions for further details.) In most cases, this simply results in a call to convert(ArgumentType, ArgumentValue).

source

for — Keyword.

for loops repeatedly evaluate the body of the loop by iterating over a sequence of values. For example:

for i in [1,4,0]
    println(i)
end

source

while — Keyword.

while loops repeatedly evaluate a conditional expression, and continues evaluating the body of the while loop so long as the expression remains true. If the condition expression is false when the while loop is first reached, the body is never evaluated. For example:

while i <= 5
    println(i)
    i += 1
end

source

Base.@time — Macro.

@time

A macro to execute an expression, printing the time it took to execute, the number of allocations, and the total number of bytes its execution caused to be allocated, before returning the value of the expression.

See also @timev, @timed, @elapsed, and @allocated.

julia> @time rand(10^6);
  0.001525 seconds (7 allocations: 7.630 MiB)

julia> @time begin
           sleep(0.3)
           1+1
       end
  0.301395 seconds (8 allocations: 336 bytes)

source

Base.@assert — Macro.

@assert cond [text]

Throw an AssertionError if cond is false. Preferred syntax for writing assertions. Message text is optionally displayed upon assertion failure.

source

Foo

@show pwd()
a = 1

pwd() = "/Users/wookyoung/.julia/v0.6/Documenter/test/examples/builds/html/lib"
1

...

isdefined(:a)

false

f(x) = 2x
g(x) = 3x

x, y = 1, 2
println(x, y)

12

type T end
t = T()

ex-3.T()

Foo

isdefined(:T)
@show typeof(T)

isdefined(:t) = true
typeof(T) = DataType
ex-3.T

x + y

3

f(2), g(2)

(4, 6)

Foo

x - y

-1

f(1), g(1)

(2, 3)

@which T()

ex-3.T() in ex-3 at none:1

Foo

a = 1
b = ans
@assert a === b

julia> nothing

julia> rand()
0.23603334566204692

julia> a = 1
1

1
julia> println(a)

julia> b = 2
2

julia> a + b
3

julia> type T
           x :: Int
           y :: Vector
       end

julia> x = T(1, [1])
##repl-#693.T(1, [1])

julia> x.y
1-element Array{Int64,1}:
 1

julia> x.x
1

julia> d = 1
1

1
julia> println(d)

Test setup function

@assert w === 5

julia> @assert w === 5
ERROR: UndefVarError: w not defined

Autodocs

AutoDocs Module

AutoDocs — Module.

AutoDocs module.

source

AutoDocs.K — Constant.

Constant K.

source

AutoDocs.T — Type.

Type T.

source

AutoDocs.f — Method.

Function f.

source

AutoDocs.@m — Macro.

Macro @m.

source

Functions, Modules, and Macros

AutoDocs.A — Module.

Module A.

source

AutoDocs.A.f — Method.

Function A.f.

source

AutoDocs.A.@m — Macro.

Macro B.@m.

source

AutoDocs.B — Module.

Module B.

source

AutoDocs.B.f — Method.

Function B.f.

source

AutoDocs.B.@m — Macro.

Macro B.@m.

source

Constants and Types

AutoDocs.A.K — Constant.

Constant A.K.

source

AutoDocs.A.T — Type.

Type B.T.

source

AutoDocs.B.K — Constant.

Constant B.K.

source

AutoDocs.B.T — Type.

Type B.T.

source

Autodocs by Page

AutoDocs.Pages.f — Method.

f from page a.jl.

Links:

ccall

while

@time(x)

T(x)

T(x, y)

f(::Integer)

f(::Any)

f(::Any, ::Any)

f(x, y, z)

[footnote]
Footnote contents. [footnote]

source

AutoDocs.Pages.f — Method.

f from page b.jl.

Links:

ccall

while

@time

T

f

f(::Any)

f(::Any, ::Any)

f(::Any, ::Any, ::Any)

source

AutoDocs.Pages.f — Method.

f from page c.jl.

Links:

ccall

while

@time

T

f

f(::Any)

f(::Any, ::Any)

f(::Any, ::Any, ::Any)

source

AutoDocs.Pages.T — Type.

T from page d.jl.

Links:

ccall

while

@time

T

f

f(x)

f(x, y)

f(::Any, ::Any, ::Any)

source

AutoDocs.Pages.T — Method.

T from page d.jl.

Links:

ccall

while

@time

T()

T(x)

T(x, y)

T(x, y, z)

f

f(x)

f(x, y)

f(::Any, ::Any, ::Any)

source

AutoDocs.Pages.T — Method.

T from page d.jl.

Links:

ccall

while

@time

T(x)

T(x, y)

T(x, y, z)

f

f(x)

f(x, y)

f(::Any, ::Any, ::Any)

source

A footnote reference [footnote].

Named docstring @refs

a normal docstring @ref link: AutoDocs.Pages.f;

a named docstring @ref link: f;

and a link with custom text: @time 1 + 2;

some invalid syntax: for i = 1:10; ....